# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is the backend API for The Mechanic Next Door (TMND).",
        "title": "TMND API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "Daniel Osariemen",
            "url": "http://github.com/thexovc",
            "email": "osazeepeter79@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/user/edit-profile": {
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Update the profile details of the currently authenticated user.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Edit user profile",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Full Name",
                        "name": "fullName",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Phone Number",
                        "name": "phoneNumber",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "Address",
                        "name": "address",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "State",
                        "name": "state",
                        "in": "formData"
                    },
                    {
                        "type": "string",
                        "description": "LGA",
                        "name": "lga",
                        "in": "formData"
                    },
                    {
                        "type": "file",
                        "description": "Profile Image",
                        "name": "profile-image",
                        "in": "formData"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "User profile updated successfully",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        },
        "/user/get-user": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "Retrieve the details of the currently authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "User"
                ],
                "summary": "Get user details",
                "responses": {
                    "200": {
                        "description": "User details retrieved successfully",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/main.Response"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.User"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/main.Response"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "main.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                },
                "statusCode": {
                    "type": "integer"
                }
            }
        },
        "models.Booking": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "bookingDate": {
                    "type": "string"
                },
                "bookingFee": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "errorMessage": {
                    "type": "string"
                },
                "estimatedPrice": {
                    "type": "number"
                },
                "id": {
                    "type": "string"
                },
                "latitude": {
                    "type": "number"
                },
                "longitude": {
                    "type": "number"
                },
                "mechanic": {
                    "$ref": "#/definitions/models.Mechanic"
                },
                "mechanicId": {
                    "type": "string"
                },
                "paymentRef": {
                    "type": "string"
                },
                "serviceDescription": {
                    "type": "string"
                },
                "serviceType": {
                    "type": "string"
                },
                "services": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Service"
                    }
                },
                "status": {
                    "$ref": "#/definitions/models.BookingStatus"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "string"
                },
                "vehicle": {
                    "description": "relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.Vehicle"
                        }
                    ]
                },
                "vehicleId": {
                    "type": "string"
                },
                "vehicleImagesUrl": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                }
            }
        },
        "models.BookingStatus": {
            "type": "string",
            "enum": [
                "pending",
                "booked",
                "cancelled",
                "completed"
            ],
            "x-enum-varnames": [
                "BookingPending",
                "BookingBooked",
                "BookingCancelled",
                "BookingCompleted"
            ]
        },
        "models.Mechanic": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "bookings": {
                    "description": "relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Booking"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "experience": {
                    "description": "years of experience",
                    "type": "integer"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "isEmailVerified": {
                    "type": "boolean"
                },
                "lga": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profileFileName": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                },
                "rating": {
                    "type": "number"
                },
                "specialty": {
                    "type": "string"
                },
                "state": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Service": {
            "type": "object",
            "properties": {
                "basePrice": {
                    "type": "number"
                },
                "bookings": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Booking"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "difficulty": {
                    "$ref": "#/definitions/models.ServiceDifficulty"
                },
                "duration": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "isAvailable": {
                    "type": "boolean"
                },
                "name": {
                    "type": "string"
                },
                "serviceCategoryId": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.ServiceDifficulty": {
            "type": "string",
            "enum": [
                "Easy",
                "Medium",
                "Hard"
            ],
            "x-enum-varnames": [
                "Easy",
                "Medium",
                "Hard"
            ]
        },
        "models.Transaction": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "currentBalance": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "mechanic": {
                    "$ref": "#/definitions/models.Mechanic"
                },
                "mechanicId": {
                    "type": "string"
                },
                "previousBalance": {
                    "type": "number"
                },
                "status": {
                    "$ref": "#/definitions/models.TrasactionStatus"
                },
                "trxRef": {
                    "type": "string"
                },
                "type": {
                    "$ref": "#/definitions/models.TransactionType"
                },
                "updatedAt": {
                    "type": "string"
                },
                "user": {
                    "description": "relationships",
                    "allOf": [
                        {
                            "$ref": "#/definitions/models.User"
                        }
                    ]
                },
                "userId": {
                    "type": "string"
                }
            }
        },
        "models.TransactionType": {
            "type": "string",
            "enum": [
                "debit",
                "credit"
            ],
            "x-enum-varnames": [
                "TransactionTypeDebit",
                "TransactionTypeCredit"
            ]
        },
        "models.TrasactionStatus": {
            "type": "string",
            "enum": [
                "pending",
                "failed",
                "success"
            ],
            "x-enum-varnames": [
                "StatusPending",
                "StatusFailed",
                "StatusSuccess"
            ]
        },
        "models.User": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "balance": {
                    "type": "number"
                },
                "bookings": {
                    "description": "relationships",
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Booking"
                    }
                },
                "createdAt": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "isEmailVerified": {
                    "type": "boolean"
                },
                "lga": {
                    "type": "string"
                },
                "phoneNumber": {
                    "type": "string"
                },
                "profileFileName": {
                    "type": "string"
                },
                "profileImageUrl": {
                    "type": "string"
                },
                "registerWithGoogle": {
                    "type": "boolean"
                },
                "state": {
                    "type": "string"
                },
                "transactions": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Transaction"
                    }
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
        "models.Vehicle": {
            "type": "object",
            "properties": {
                "brand": {
                    "type": "string"
                },
                "createdAt": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "model": {
                    "type": "integer"
                },
                "size": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
    }
}